hack/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.34.0
python-multipart==0.0.9
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
